task 4 update LoginView
